feat(styleguide) Add style guide & theme-ability (#75)
Creates a shared stylesheet & kong-theme based on the newest styleguide.

This also introduces Vuepress docs, which acts as a living style guide for non
component styles (spacing, type, etc).

Inside the /styles directory is a number of individual scss stylesheets and a
variables sheet. By running yarn build:styles node-sass will compile into a
single style.css file to be released with @kongponents/styles.

_variables.scss: This sheet sets the base values for colors, sizing, text size,
and font families. \_typography.scss: This sheet loops over the text-size values
and creates classes for each for both regular and mono type faces. \_utils.scss:
This sheet loops over all spacings and creates utility classes (pt-2, mb-4,
etc). It also creates a few generic utility classes for display, truncating,
etc. styles.scss: This sheet imports the above and generates :root {} variables
that can be used in applications & components.

Theming: Kongponents will be updated to use a theme for their styles which will
pull from the base styles.css and theme root vars. This could be provided via a
javascript file, a stylesheet or inline in a project. The idea is to create
theme of vars which use the base styles for Kongponents to use.

Changelog:

- feat(Vuepress) Add vuepress for styleguide documentation

- docs(styleguide) Add root stylesheet to storybook and vuepress
- docs(styleguide) Add colors page and swatch
- docs(styleguide) Add type page and textblock component
- docs(styleguide) Update styling of docs
- docs(styleguide) add roboto mono to storybook/vuepress
- docs(type) document utility classes

- chore(styles) move styles directory to packages