fix(*): fix deps and build process
Update build process to copy necessary config files in a separate step,
leave them in place until build:cli finishes, then remove, as other steps
in the build process will fail if postcss.config.js files are not
left in place.

Update @vue/composition-api to be a peerDependency.

Add docs info on KTable and KCardCatalog for how to use with
Composition API.