fix(kbutton): not considering visually-hidden text
Replaces the `padding-right` declaration on a button’s icon with a `gap` property on the button itself which allows adding visually-hidden text to an icon button (this would otherwise have an incorrect gap to the right). This works because visually-hidden text won’t participate as a flex or grid item so no gap will be rendered.