refactor(kcodeblock): tweaks actions layout
Reworks the layout of the items in the code block’s action bar to avoid jumping content in almost all sizes of the code block. The error message is now displayed absolutely positioned right below the search box as to not take up physical space in the actions bar. The result message is now positioned to the right of the search box and always takes up a fixed amount of space to avoid shifting around other elements. From about 280 pixels of width, the actions bar will wrap to 3 rows of elements while still avoiding layout shifts.