fix(kcodeblock): a couple of issues (#846)
* fix(kcodeblock): handling small resolutions poorly

Addresses a few issues with the rendering of `KCodeBlock` in smaller resolutions. The actions bar items are now wrapping appropriately and the search input no longer extends beyond the available size.

* fix(kcodeblock): shortcuts using Alt not working on macOS

Fixes a bug where the default shortcuts involving the Alt key weren’t working on macOS because Options+letter commonly results in a differently typed character. The shortcut mechanism now checks the `event.code` property instead of `event.key` to do this more reliably.

* fix(kcodeblock): not preventing default action for some shortcuts

Fixes an issue with not correctly preventing the default action for some shortcuts (e.g. `Alt+F` opening a Chrome menu by default).

* refactor(kcodeblock): tweaks actions layout

Reworks the layout of the items in the code block’s action bar to avoid jumping content in almost all sizes of the code block. The error message is now displayed absolutely positioned right below the search box as to not take up physical space in the actions bar. The result message is now positioned to the right of the search box and always takes up a fixed amount of space to avoid shifting around other elements. From about 280 pixels of width, the actions bar will wrap to 3 rows of elements while still avoiding layout shifts.