fix(kcodeblock): code-block-render event being emitted too soon
Fixes a small issue with the `code-block-render` event being emitted before re-rendering is actually done. This is only an issue **when changing the `props.code`** as the event is already delayed by one Vue tick when turning off filter mode and doesn’t need to be delayed when emitting the event as part of the `mounted` life cycle hook.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>