fix(kpop): stop document click event propagation only when the event is about to hide the popper
This fixes the issue where other document click event listeners were not being fired because
the stopPropagation() was always being called no matter what is the actual action of the click.

Most of the UI libraries use document-bound click listeners for execution optimization reasons
and this behavior caused them to not work anymore.

stopPropagation() should only be called when the click event is hiding the poppep.