feat(ktable): loading state [khcp-7162] (#1338)
Update to tie the loading state of `KTable` to SWRV's `PENDING` state instead of `VALIDATING` which fires even when pulling from cache. Also adds a new prop `cacheIdentifier` which is used to identify the table's entry in the cache rather than using a randomly generated identifier that prevents the ability to use caching. Lastly, this also fixes a bug in `useSwrvState` that caused the composable to incorrectly determine whether or not the response has data due to different response structures.
For [KHCP-7162](https://konghq.atlassian.net/browse/KHCP-7162).