fix(kpagination): kpopAttrs not being defined but used (#1427)
* chore: ensures pagesVisible is list of number

Changes `getVisiblePages` to return `[]` when `props.disablePageJump` is true instead of `0` allowing the function’s return type to be `number[]` instead of `number | number[]` which is likely unintentional seeing as `pageVisible` is used in a `v-for` directive. This silences the TypeScript error shown in the template related to this.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

* fix(kpagination): kpopAttrs not being defined but used

Adds `kpopAttrs` again which was lost in a previous refactoring. This ensures `placement: top` is passed to `KSelect` and silences the Vue warning appearing during development.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>

---------

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>