fix(kslideout): title styles
Addresses a number of style issues of the title bar in order to achieve a better default appearance.

Removes the margins added to the title, before title, and after title containers and instead adds a padding to the header making the title bar appear flush with the content.

Reduces the gap size between title, before title, and after title to allow for more flexible usage scenarios.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>