fix(kpagination): ensure currentPage is highlighted (#2307)
If your current page is not in the set of page buttons that KPagination defaults to (e.g. [1] [2] [3] [...] [20] with a current page of `7`). This fix ensures that the correct set of page buttons is displayed plus highlights the button that corresponds to the current page.

I also added an example to the sandbox which I used to figure this out.

FYI: Our use case is that we always pass bookmarkable `?page=7` parameters to our pages, previous to this fix a '7 page button' would not be shown or highlighted, also see https://github.com/kumahq/kuma-gui/issues/2769

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>